+++ components/maker/index.js
@@ -1,10 +1,9 @@
-import { SidebarButton } from './styled'
-import PaymentDetails from './payment-details'
-import ParticipationAgreement from './participation-agreement'
+import PaymentDetails from './payment'
+import ParticipationAgreement from './legal'
 import KYC from './kyc'
 import Status from './status'
 import Notification from './notification'
-import Info from './info'
+import { SidebarButton } from './styled'
 
 
 export default {
@@ -13,6 +12,5 @@ export default {
   ParticipationAgreement,
   KYC,
   Status,
-  Notification,
-  Info
+  Notification
 }
